Add new testing module final.
These tests:

  - maven-archetype-plugin/src/it/create-from-project
  - DefaultArchetypeCreatorTest.java
  - ArchetyperRoundtripTest.java
  - ArchetyperRoundtripWithProxyTest.java
  - RoundtripMultiModuleTest.java

require the presence of org.apache.maven.plugins:maven-archetype-plugin:${project.version}.
That's why we can't keep them in the archetype-common or maven-archetype-plugin modules, and thus
the reason to introduce a new module in the archetype-testing module. The new module archetype-final
depends on both archetype-common and maven-archetype-plugin so it gets executed last and contains
tests that require everything else to be built before that.