[ARCHETYPE-605] Allow .gitignore file in archetype resources
This change allows to create archetypes with files like .gitignore
that are excluded by default. It adds the parameter
"addDefaultExcludes" to the jar mojo. The parameter can be set to
"false" to allow including default files like .gitignore.
The default is "true" (exclude default files) so the plugin behaves
the same as before when the parameter is not set. The parameter is
named and behaves the same as the one in the plugin
"maven-resources-plugin" which usually also needs to be set when
someone wants to include one of the files excluded by default.